reorganized photo page files
