+++ src/photos/pages/index/PhotoIndexPage.tsx
@@ -32,8 +32,8 @@ export const PhotoIndexPage: NextPage<Props> = ({ folderName, images }: Props) =
         <>
             <Head>
                 <title>{folderName}</title>
-                {/* <meta property="og:image" content="https://nextjsconf-pics.vercel.app/og-image.png" />
-                <meta name="twitter:image" content="https://nextjsconf-pics.vercel.app/og-image.png" /> */}
+                <meta property="og:image" content="https://nextjsconf-pics.vercel.app/og-image.png" />
+                <meta name="twitter:image" content="https://nextjsconf-pics.vercel.app/og-image.png" />
             </Head>
             <main className="mx-auto max-w-[1960px] p-4">
                 {photoId && (
