+++ src/photos/pages/index/PhotoIndexPageGetStaticProps.ts
